big int for sale item type

+++ api/src/migrations/20210325212836_recreate_sale_offers.ts
@@ -3,7 +3,7 @@ import * as Knex from "knex";
 export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTable("sale_offers", async (table) => {
     table.integer("sale_item_id").primary();
-    table.integer("sale_item_type").notNullable();
+    table.bigInteger("sale_item_type").notNullable();
     table.text("sale_item_owner");
     table.decimal("sale_price");
     table.text("transaction_id");
